Modify WebKit automatic scroller
Declare readonly -documentView to make it easier to access the document view of the frame view without having to delcare the frame view every time the document view is accessed.

This commit also modifies the behavior of -viewingBottom. It no longer relies solely on _userScrolled. It will check the actual position of the view if _userScrolled is true.