Modify WebKit automatic scroller
Add -resetScrollerPosition and -resetScrollerPositionTo: which allows an outside class to modify whether the automatic scroller will restore position to the bottom when the view becomes visible.

This is used by the jump to line logic so that if the view is no longer scrolled to the bottom once a jump completes, the automatic scroller is aware of this fact without relying on notifications.