Modify ICLPayload
There is no easy way of rendering templates from XPC Services, so we modify ICLPayload to accommodate this limitation.

Instead of the payload containing the rendered HTML, we set an entrypoint property which is the name of a JavaScript function and an accompanying entrypointPayload dictionary for context.

The JavaScript is then called in the context of the WebView which can easily render templates. (not yet implemented)