Workaround for strange bug
A copy of the app downloaded from the App Store returns
an empty array of products for some reason when the receipt
file is missing. I cannot replicate this behavior in the
developer sandbox which suggests its an issue specific to
the signing certificate the App Store replaces the code
signature with. To prevent crashes, we check if the list
of products contain identifiers we always expect to exist.