Stop using hard coded layout constraints in almost all locations
Many constraints are changed dynamically from 0 back to their original value. The original value used to be kept in a macro. Now, a custom property named -archivedConstant is added to NSLayoutConstraint which stores whatever value of -constant is at the time -archiveConstant is called. We can then call -restoreArchivedConstant to restore it when we want the default back.

This allows us to make changes in interface builder and have that reflected in code automatically.

There is one or two places where this isn't possible because there is more than one default depending on layout of other elements. This patch still covers most cases.