Add TVCAlert and TDCAlert
- Added TVCAlert
TVCAlert acts as a non-blocking substitute to NSAlert which can be used to show messages that aren't important.

- Added TDCAlert
TDCAlert is an interface for TDCAlert that abstracts away most of the complexity.

TDCAlert also replaces TLOPopupPrompts as interface for regular NSAlert modal alerts and sheet alerts.

TLOPopupPrompts interacts with TDCAlert on behalf of code that still uses the former. TLOPopupPrompts is formally marked as deprecated and will eventually be removed from the codebase.

These commit also modifies -[THOPluginManager extrasInstallerCheckForUpdates] to show a single update alert for all available updates instead of one for each addon. The alert also uses the new non-blocking version.