Modify behavior of nickname changes on private messages
Setup:
• A private message exists for user GuestA and GuestB.
• GuestB is offline which means GuestA can change their nickname to GuestB.
• GuestA changes their nickname to GuestB.

Previous behavior:
The private message window for GuestB is destroyed, including its saved scrollback. GuestA's window is then renamed to GuestB.

New behavior:
If a private message exists as the new name, then nothing happens. The window for the old nickname is not renamed. The window for the new nickname is not destroyed.