Add -contentIsFile flag
Returns YES when content is a link to a file such as image or video.

Unlike other properties, enable this on a per-module basis instead of letting it be inherited by one of the root classes. In the future, whether a root class defines this property will fade out of memory. It's best to assume it's not.