Add -okOrErrorForTextField: and -okOrErrorForComboBox:
These two methods act as a shorthand to help hosts validate an element quicker. They are supposed to be called within -ok: before the sheet is dismissed.

If the return value is YES, then the value is valid and -ok: can proceed.

If the return value is NO, then the value is invalid, the error popover will automatically present itself on behalf of the host, and -ok: cannot proceed.