ADD TVCAppearance
TVCAppearance is an object, that when subclassed, adds a very powerful foundation for loading properties for themeing a user interface.

Given a dictionary, the top level keys of the dictionary are considered the available appearances. The object is allocated given a specific appearance name and that is loaded from the dictionary. Support for inherting from other appearances is supported to allow for fine grain modifications under specific conditions. e.g. graphite on OS X Mavericks

The object also provides an interface to easy extra common appearance properties such as colors, fonts, images, and measurements. The accessors for these resources automatically pick a 1x or 2x value depending on whether high resolution mode (retina) is enabled.