Do not use TVCMainWindowSidebarYosemiteSmoothTextFieldCell on Mojave
Created a separate view for Mojave which doesn't use the smooth text field because that's no longer needed thanks to improvements Apple has made.
We can't just disable it in the drawing code by calling super because that seems to create weird effects that are eliminated by removing it. Maybe with time, I could have learned why that is, but this solution is quicker.