Modify TVCMainWindowChannelView
The appearance of the channel view is dependent on the background color of the theme, not that of the app or system. This commit seperates appearance logic into the object itself, making itself contained. When view moves to the window, it adds an observer to the theme settings to know when theme setting for -underlyingWindowColorIsDark changes.