Transition core to use random IDs for localization in place of integer
If you're wondering why I don't follow the pattern of "<localization>" = "<localization>" which is more commonly accepted: it's because random keys make code cleaner. In my opinion. It also makes it easier to access different tables. TXTLS(), which is the function that returns strings, accepts first half of key as the table. For example "IRC[somekey]" translates to: lookup "somekey" in table "IRC" — There are others, but that is the most notable.

This commit also moves some strings to different files including the addition of a new file named CommonErrors.strings.