Rewrite connection manager in Swift
The entire connection manager service, except for the main entry point, has been rewritten in Swift. The connection interface is now modular by using subclasses. The reason it was written in Swift is because new Network.framework which there are plans to support doesn't have an Objective-C interface. There is only a plain C interface or Swift.