Replace "Prefer IPv4" preference
Because modern sockets do not support setting a preference for address type, we now make the address type explicit. User can pick to connect to only IPv4, only IPv6, or have it automatically chosen.

For users that had the old preference enabled, they will automatically be migrated to IPv4 only. Additionally, they are shown a warning each time they connect to inform them of the change in behavior. The warning disappears if they change the new preference, which is located in Network Socket section of Server Properties.