Automatically enable modern sockets
No reason to have a per-connection flag to enable or disable modern sockets.

If user is on an a version of macOS that supports it and they have none or automatic proxy type configured, then we use it.

Just in case there is some fallout for some unforeseen problem, a hidden preference will remain that can be used to turn it off app wide.