Always assign nil to self.appearance when appKitAppearanceTarget is not applicable
When a specific appearance is used, self.appearance may be assigned an NSAppearance object. When system appearance is used, the target specifies that self.appearance shouldn't contain a value so that it can be inherited, but because we were only setting self.appearance when appKitAppearanceTarget is applicable; the NSAppearance object already set was never unset.