Attempt to resolve "Close Query" keyboard shortcut issue
Only one menu item can have a keyboard shortcut which means because "Delete Channel" and "Close Query" are competiting for the same, there must be a way to determine which wins. Previous solution was to archive the keyboard shortcut set and restore it. This proved to be prone to failure. New solution is much simpler. We simply unassign the submenu when the parent menu is hidden so that the only time the keyboard shortcut is in the main menu is when it's the one we want to use.