Add auto join delay timer preference
On a certain server, even though we received message from NickServ that identification was successful, joining channels failed because of lag between NickServ and the IRCd. This commit adds a new preference to the Flood Control section of Preferences which controls delay before auto join is performed. It defaults to 0.0 (no delay). It can be configured up to 10.0 seconds.

To accommodate this new feature, the current auto join timer has been renamed to auto join next join timer.