Add window in which custom CSS rules can be defined.
Because the rules added in this window are completely local to the client and has no influence from the outside, there is absolutely no sanitation performed. This allows for open abuse of the system by escaping from the <style> element to append JavaScript or anything else without making a local copy of the style.