Fix text field truncation causing crash
I do not know what I did at the time but switching to using -enumerateAttributesInRange:options:usingBlock: to perform truncation was an absolute mad idea because we need to be able to recalculate the segment range on the fly using the composed character length. This commit reverts to manually calculating segment ranges.