Stop shipping with Growl support
All code related to Growl is no longer shipped by default through the use of the TEXTUAL_BUILT_WITH_GROWL_SDK_ENABLED macro. This macro can be set to the value "1" to enable the code.

Growl.framework was built against the 10.8 SDK which is not compatible with a hardened runtime. I do not have the time or resources to research alternate solutions at this time, but hope another person may in my stead.