Refactor member list
All member list logic inside IRCChannel has been abstracted away into IRCChannelMemberList.
IRCChannelMemberListController, which is an array controller, is now bound to the list in the main window to act as the data source.

The controller is weakly linked in IRCChannelMemberList when the channel is selected. This allows IRCChannelMemberList to maintain the UI in addition to its internal list.

The content of the controller shares 1:1 mutable array to IRCChannelMemberList's internal list. This simplifies modifications to the UI as we know with certainty that the index of the member in the internal list will be that of the member in the UI.

The methods declared by IRCChannel still exists, as a proxy, to an instance of IRCChannelMemberList strong referenced by the -memberInfo property of the channel.