Fix regression in TDCAlert introduced by 89ac727
I modified method to return TDCAlertResponse type instead of a boolean. This is enum with three values: 1000, 1001, 1002. It is declared as an integer. Compiler also treats YES (objective-c equivalent of TRUE) as an integer (1) so compiler did not notify me I was returning an incorrect value when I forgot to modify one return since it was in fact returning an integer. But because YES (1) did not equal one of the expected results, it threw off the logic.