Use NSInvocation rather than raw objc_msgSend calls
On Intel-based Macs, objc_msgSend is variadic, as the variadic calling
convention in use matches the non-variadic one. On Arm-based Macs,
objc_msgSend is defined as void objc_msgSend(void), so any attempt to
pass arguments to it gives an error. If you want to call objc_msgSend on
Arm-based Macs you are thus required to cast a function pointer to the
desired type (and on Intel-based Macs this is the same as the variadic
calling convention). However, directly calling objc_mstSend still relies
on implementation details of the Objective-C runtime which, whilst true,
is unnecessary, so instead use the higher-level NSInvocation, even if it
is a little verbose.