Mac model list spring cleaning
All Macs that are not compatible with macOS 10.13 or greater have been removed from the list. A lot of other redundant information, such as specifying "retina", "M1", etc, for some models, has been removed. This information can be derived from other parts of the system information command. The date on which a model has been released no longer have a prefix to determine when in the year the release occurred if not more than one release occurred in the year.